# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Streammux Job Management API",
        "description": "REST API for job definitions and lifecycle commands. The service validates payloads, maintains a Kafka-projected read model, and publishes definitions, commands, and events to Kafka.",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0-SNAPSHOT"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
        "/jobs/{jobId}": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get job",
                "operationId": "get",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job definition",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobDefinition"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobDefinition"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Update job",
                "description": "Increments version, validates payload, and publishes the updated definition.",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JobDefinition"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Job updated",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobDefinition"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobDefinition"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Delete job",
                "description": "Marks the job deleted, publishes commands/events, and removes it from the local read model.",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Delete accepted"
                    },
                    "404": {
                        "description": "Job not found"
                    }
                }
            }
        },
        "/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "List jobs",
                "operationId": "list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/JobDefinition"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Create job",
                "description": "Registers a new job and publishes the definition to Kafka.",
                "operationId": "create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JobDefinition"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Job created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobDefinition"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Job id already exists",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobDefinition"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/jobs/{jobId}/resume": {
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Resume job",
                "operationId": "resume",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Resume command accepted"
                    },
                    "404": {
                        "description": "Job not found"
                    }
                }
            }
        },
        "/jobs/{jobId}/restart": {
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Restart job",
                "operationId": "restart",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Restart command accepted"
                    },
                    "404": {
                        "description": "Job not found"
                    }
                }
            }
        },
        "/jobs/{jobId}/rename": {
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Rename job",
                "description": "Copies the current definition under the new jobId key, tombstones the old key with desiredState=DELETED, and emits paired CREATED/DELETED events with rename attributes. All other fields (jobType, desiredState, config, labels, tags, etc.) are copied from the existing definition; use PUT /jobs/{jobId} first if you need to change them at the same time. Runtime state (status, lease) and version history do not carry over; any running runner restarts under the new id.",
                "operationId": "rename",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Current job identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RenameJobRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Job renamed; returns the new definition",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobDefinition"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid or missing newJobId (blank, or equal to path jobId)",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobDefinition"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobDefinition"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "newJobId already exists",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobDefinition"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/jobs/{jobId}/pause": {
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Pause job",
                "operationId": "pause",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Pause command accepted"
                    },
                    "404": {
                        "description": "Job not found"
                    }
                }
            }
        },
        "/jobs/validate": {
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Validate job",
                "description": "Runs the same validator used by create/update/rename without persisting or publishing anything. Returns 200 with {\"valid\":true} when the definition passes and 400 VALIDATION_ERROR (matching a real save) when it does not.",
                "operationId": "validate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JobDefinition"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Definition is valid",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidateJobResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Definition failed validation; body has the VALIDATION_ERROR envelope",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidateJobResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/jobs/{jobId}/status": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Runtime status",
                "description": "Returns JSON when status exists for this job; empty body with HTTP 200 when the read model has no status yet.",
                "operationId": "status",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobRuntimeStatus"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/jobs/{jobId}/lease": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Lease",
                "description": "Returns JSON when a lease exists for this job; empty body with HTTP 200 when none is projected yet.",
                "operationId": "lease",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobLease"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/jobs/{jobId}/events": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Job events",
                "description": "Audit-style events for the job from the read model (may be empty).",
                "operationId": "events",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/JobEvent"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "JobDefinition": {
                "type": "object",
                "properties": {
                    "jobId": {
                        "type": "string"
                    },
                    "jobVersion": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "jobType": {
                        "type": "string",
                        "enum": [
                            "ROUTE_APP",
                            "RANDOM_SAMPLER",
                            "ALARMS_TO_ZTR"
                        ]
                    },
                    "desiredState": {
                        "type": "string",
                        "enum": [
                            "ACTIVE",
                            "PAUSED",
                            "DRAINING",
                            "DELETED"
                        ]
                    },
                    "priority": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "siteAffinity": {
                        "type": "string"
                    },
                    "leasePolicy": {
                        "$ref": "#/components/schemas/LeasePolicy"
                    },
                    "parallelism": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "routeAppConfig": {
                        "$ref": "#/components/schemas/RouteAppConfig"
                    },
                    "randomSamplerConfig": {
                        "$ref": "#/components/schemas/RandomSamplerConfig"
                    },
                    "alarmsToZtrConfig": {
                        "$ref": "#/components/schemas/AlarmsToZtrConfig"
                    },
                    "labels": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedBy": {
                        "type": "string"
                    }
                }
            },
            "LeasePolicy": {
                "type": "object",
                "properties": {
                    "heartbeatIntervalSeconds": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "leaseDurationSeconds": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "claimBackoffMillis": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "allowFailover": {
                        "type": "boolean"
                    }
                }
            },
...
            "RenameJobRequest": {
                "type": "object",
                "description": "Request body for POST /jobs/{jobId}/rename",
                "properties": {
                    "newJobId": {
                        "type": "string",
                        "description": "New job id under which the definition should be published",
                        "example": "route-poc-1-renamed"
                    }
                }
            },
            "ValidateJobResponse": {
                "type": "object",
                "description": "Response body for POST /jobs/validate when the definition passes validation",
                "properties": {
                    "valid": {
                        "type": "boolean",
                        "description": "Always true; validation failures return HTTP 400 with a VALIDATION_ERROR envelope instead",
                        "example": true
                    }
                }
            },
            "JobEvent": {
                "type": "object",
                "properties": {
                    "eventId": {
                        "type": "string"
                    },
                    "jobId": {
                        "type": "string"
                    },
                    "jobVersion": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "eventType": {
                        "type": "string",
                        "enum": [
                            "SESSION",
                            "CREATED",
                            "UPDATED",
                            "CLAIMED",
                            "STARTED",
                            "HEARTBEAT",
                            "PAUSED",
                            "RESUMED",
                            "FAILED",
                            "RELEASED",
                            "STOPPED",
                            "DELETED"
                        ]
                    },
                    "eventTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "siteId": {
                        "type": "string"
                    },
                    "instanceId": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "attributes": {
                        "type": "object",
                        "additionalProperties": {}
                    },
                    "actor": {
                        "type": "string"
                    }
                }
            },
            "JobRuntimeStatus": {
                "type": "object",
                "properties": {
                    "jobId": {
                        "type": "string"
                    },
                    "jobVersion": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "state": {
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "STARTING",
                            "RUNNING",
                            "PAUSED",
                            "DEGRADED",
                            "FAILED",
                            "STOPPED"
                        ]
                    },
                    "health": {
                        "type": "string",
                        "enum": [
                            "UNKNOWN",
                            "HEALTHY",
                            "DEGRADED",
                            "UNHEALTHY"
                        ]
                    },
                    "lastHeartbeatAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "workerMetadata": {
                        "$ref": "#/components/schemas/WorkerMetadata"
                    },
                    "failureReason": {
                        "type": "string"
                    },
                    "lagMetrics": {
                        "$ref": "#/components/schemas/LagMetrics"
                    }
                }
            },
...
            "JobLease": {
                "type": "object",
                "properties": {
                    "jobId": {
                        "type": "string"
                    },
                    "jobVersion": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "leaseOwnerSite": {
                        "type": "string"
                    },
                    "leaseOwnerInstance": {
                        "type": "string"
                    },
                    "leaseEpoch": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "CLAIMED",
                            "RUNNING",
                            "RELEASING",
                            "LOST",
                            "EXPIRED"
                        ]
                    },
                    "leaseExpiresAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastHeartbeatAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        }
    }
}